requirements.txt
fastapi
uvicorn
sqlalchemy
openai
python-dotenv
jinja2
psycopg2-binary
pydantic
pandas 